@@ -1,6 +1,6 @@
 package com.baeldung.spring.reactive.performance;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
