@@ -3,7 +3,11 @@ package com.baeldung.java9.rangedates;
 import java.util.Collection;
 import java.util.Date;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 public class DatesCollectionIteration {
+    private static final Logger log = LoggerFactory.getLogger(DatesCollectionIteration.class);
 
     public void iteratingRangeOfDatesJava7(Collection<Date> dates) {
 
@@ -18,7 +22,7 @@ public class DatesCollectionIteration {
     }
 
     private void processDate(Date date) {
-        System.out.println(date);
+        log.debug(date.toString());
     }
 
 }
